Remove env variable used as FF
