+++ activiti-core/activiti-api-impl/activiti-api-process-runtime-impl/src/test/resources/jsonPatch/jsonPatch-in-mapping-output.json
@@ -10,9 +10,43 @@
           "required": true,
           "value": "John"
         },
-        "variable-person-id": {
-          "id": "variable-person-id",
-          "name": "process_variable_person",
+        "variable-person-simple-cases-id": {
+          "id": "variable-person-simple-cases-id",
+          "name": "process_variable_person_simple_cases",
+          "type": "json",
+          "required": false,
+          "value": {
+            "firstname": "Bob",
+            "addresses": [
+              {
+                "street": "123 Main St"
+              },
+              {
+                "street": "456 Elm St"
+              }
+            ]
+          }
+        },
+        "variable-person-array-cases-id": {
+          "id": "variable-person-array-cases-id",
+          "name": "process_variable_person_array_cases",
+          "type": "json",
+          "required": false,
+          "value": {
+            "firstname": "Bob",
+            "addresses": [
+              {
+                "street": "123 Main St"
+              },
+              {
+                "street": "456 Elm St"
+              }
+            ]
+          }
+        },
+        "variable-person-variable-cases-id": {
+          "id": "variable-person-variable-cases-id",
+          "name": "process_variable_person_variable_cases",
           "type": "json",
           "required": false,
           "value": {
@@ -57,6 +91,34 @@
               "street": "Ha-Ha Road"
             }
           }
+        },
+        "variable-property-path-id": {
+          "id": "variable-property-path-id",
+          "name": "process_variable_property_path",
+          "type": "string",
+          "required": true,
+          "value": "propertyFromVariable"
+        },
+        "variable-index-1": {
+          "id": "variable-index-1-id",
+          "name": "process_variable_index_1",
+          "type": "integer",
+          "required": true,
+          "value": 1
+        },
+        "variable-index-2": {
+          "id": "variable-index-2-id",
+          "name": "process_variable_index_2",
+          "type": "integer",
+          "required": true,
+          "value": 2
+        },
+        "variable-street-name-id": {
+          "id": "variable-name-id",
+          "name": "process_variable_street_name",
+          "type": "string",
+          "required": true,
+          "value": "Street Name"
         }
       },
       "mappings": {
@@ -72,14 +134,19 @@
             }
           },
           "outputs": {
-            "process_variable_person": {
+            "process_variable_person_simple_cases": {
               "type": "JSONPATCH",
               "value": [
                 {
                   "op": "add",
                   "path": "/lastname",
                   "value": "${process_variable_lastname}"
-                },
+                }
+              ]
+            },
+            "process_variable_person_array_cases": {
+              "type": "JSONPATCH",
+              "value": [
                 {
                   "op": "add",
                   "path": "/addresses/1/street",
@@ -87,12 +154,37 @@
                 },
                 {
                   "op": "add",
-                  "path": "/addresses/1/new-field",
-                  "value": "${process_variable_name}"
+                  "path": "/addresses/1/new-street-field",
+                  "value": "${process_variable_street_name}"
                 },
                 {
                   "op": "add",
                   "path": "/addresses/2",
+                  "value": "${process_variable_address}"
+                }
+              ]
+            },
+            "process_variable_person_variable_cases": {
+              "type": "JSONPATCH",
+              "value": [
+                {
+                  "op": "add",
+                  "path": "/${process_variable_property_path}",
+                  "value": "${process_variable_lastname}"
+                },
+                {
+                  "op": "add",
+                  "path": "/addresses/0/${process_variable_property_path}",
+                  "value": "${process_variable_street_name}"
+                },
+                {
+                  "op": "add",
+                  "path": "/addresses/${process_variable_index_1}/${process_variable_property_path}",
+                  "value": "${process_variable_street_name}"
+                },
+                {
+                  "op": "add",
+                  "path": "/addresses/${process_variable_index_2}",
                   "value": "${process_variable_address}"
                 }
               ]
